renaming EiffelEvents.RabbitMq.Client to be EiffelEvents.Clients.RabbitMq in code files
