+++ examples/EiffelClient.BasicPublisher/EiffelClient.BasicPublisher.csproj
@@ -10,7 +10,7 @@
 
     <ItemGroup>
       <ProjectReference Include="..\..\src\EiffelEvents.Net\EiffelEvents.Net.csproj" />
-      <ProjectReference Include="..\..\src\EiffelEvents.RabbitMq.Client\EiffelEvents.RabbitMq.Client.csproj" />
+      <ProjectReference Include="..\..\src\EiffelEvents.Clients.RabbitMq\EiffelEvents.Clients.RabbitMq.csproj" />
     </ItemGroup>
 
 </Project>
