+++ examples/EiffelClient.PublisherOne/Models/Edition-Paris/Client.cs
@@ -14,7 +14,7 @@
 
 using EiffelEvents.Net.Clients;
 using EiffelEvents.Net.Events.Edition_Paris;
-using EiffelEvents.RabbitMq.Client;
+using EiffelEvents.Clients.RabbitMq;
 
 namespace EiffelClient.PublisherOne.Models.Edition_Paris
 {
