+++ examples/EiffelClient.SubscriberOne/EiffelClient.SubscriberOne.csproj
@@ -7,7 +7,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\EiffelEvents.Net\EiffelEvents.Net.csproj" />
-    <ProjectReference Include="..\..\src\EiffelEvents.RabbitMq.Client\EiffelEvents.RabbitMq.Client.csproj" />
+    <ProjectReference Include="..\..\src\EiffelEvents.Clients.RabbitMq\EiffelEvents.Clients.RabbitMq.csproj" />
   </ItemGroup>
 
   <ItemGroup>
