+++ examples/EiffelClient.SubscriberOne/Program.cs
@@ -18,8 +18,8 @@ using EiffelEvents.Net.Clients;
 using EiffelEvents.Net.Clients.Validation;
 using EiffelEvents.Net.Events.Core;
 using EiffelEvents.Net.Events.Edition_Lyon;
-using EiffelEvents.RabbitMq.Client;
-using EiffelEvents.RabbitMq.Client.Config;
+using EiffelEvents.Clients.RabbitMq;
+using EiffelEvents.Clients.RabbitMq.Config;
 using FluentResults;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
