+++ examples/EiffelClient.SubscriberTwo/Program.cs
@@ -17,8 +17,8 @@ using System.Threading;
 using EiffelEvents.Net.Clients;
 using EiffelEvents.Net.Events.Core;
 using EiffelEvents.Net.Events.Edition_Lyon;
-using EiffelEvents.RabbitMq.Client;
-using EiffelEvents.RabbitMq.Client.Config;
+using EiffelEvents.Clients.RabbitMq;
+using EiffelEvents.Clients.RabbitMq.Config;
 using FluentResults;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
@@ -40,8 +40,8 @@ namespace EiffelClient.SubscriberTwo
             Console.WriteLine("Started ....");
 
             // Subscribe to events
-            _subscriptionId = _client.Subscribe<EiffelActivityCanceledEvent>(_queueIdentifier, GeneralHandleEvent);
-            Console.WriteLine($"Subscription done to event {nameof(EiffelActivityCanceledEvent)} !");
+            _subscriptionId = _client.Subscribe<EiffelActivityTriggeredEvent>(_queueIdentifier, GeneralHandleEvent);
+            Console.WriteLine($"Subscription done to event {nameof(EiffelActivityTriggeredEvent)} !");
 
             while (true)
             {
