+++ src/EiffelEvents.Clients.RabbitMq/EiffelEvents.Clients.RabbitMq.csproj
@@ -5,6 +5,7 @@
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>disable</Nullable>
         <Version>1.0.1</Version>
+        <RootNamespace>EiffelEvents.Clients.RabbitMq</RootNamespace>
     </PropertyGroup>
     
     <ItemGroup>
